\name{us}
\docType{data}
\alias{us}
\title{Allele counts for six US subpopulations.}
\description{
  9 STR loci were typed in sample populations of African Americans,
  U.S. Caucasians, Hispanics, Bahamians, Jamaicans, and Trinidadians.
}
\format{A list of tables with allele counts.}
\source{http://www.fbi.gov/hq/lab/fsc/backissu/july1999/budowle.htm}
\references{
Budowle, B., Moretti, T. R., Baumstark, A. L., Defenbaugh, D. A., and
Keys, K. M. Population data on the thirteen CODIS core short tandem
repeat loci in African Americans, U.S. Caucasians, Hispanics, Bahamians,
Jamaicans, and Trinidadians, Journal of Forensic Sciences. 1999.
}
\keyword{datasets}